session_catalog: carve rows_usage.rs (pure move)
Session usage accounting and list-row assembly: per-backend usage
parsing, repricing, sort keys, source-preserving truncation — with
tests, verbatim. SessionUsage fields widened; the shared total_usage
test fixture rides along, so its tests mod is pub(crate) and caches.rs
imports it from there. Purity-checked; tests conserved at 104.
