+++ src/bin/caller/web_gateway/session_catalog/caches.rs
@@ -812,7 +812,7 @@ pub(crate) fn file_size(path: &Path) -> u64 {
 mod tests {
     use super::*;
     use tempfile::tempdir;
-    use crate::web_gateway::session_catalog::tests::{total_usage};
+    use crate::web_gateway::session_catalog::rows_usage::tests::total_usage;
 
     fn persisted_test_key(extra: &str) -> SessionListCacheKey {
         SessionListCacheKey {
